+++ src/Enums/BillingSchedule/PhaseFrequency.php
@@ -13,5 +13,7 @@ enum PhaseFrequency: string
     case WEEKLY = 'weekly';
     case BIWEEKLY = 'biweekly';
     case MONTHLY = 'monthly';
+    case EVERY_3_MONTHS = 'every_3_months';
+    case EVERY_6_MONTHS = 'every_6_months';
     case YEARLY = 'yearly';
 }