+++ src/Organization/ReceiptPrefix.php
@@ -13,9 +13,13 @@ use Enlivy\ApiResource;
  * @property string $object
  * @property string $organization_id
  * @property string|null $name
+ * @property string|null $description
  * @property string|null $prefix
  * @property string|null $suffix
  * @property int $current_number
+ * @property bool $reset_yearly
+ * @property int|null $counter_year
+ * @property string $formatted_number
  * @property bool $has_year_prefix
  * @property bool $has_month_prefix
  * @property bool $has_day_prefix